Try using Node
On Bun, the HTML dump insertion is extremely slow (<100 rows per second
with frequent pauses), and exhibits a behavior where it gets slower over
time. My guess is that this has something to do with its text
decode (bytes -> string) performance.

With Node*, the pauses are still a thing, but they are much less common,
and in between pauses we're looking at like 2000 rows per second. The
XML dump insertion is about 50% slower (bun:sqlite is really fast), but
this is worth it.

*works by the next commit, but I want to keep this attempt with jiti in.