Use SvelteKit instead
SvelteKit's router is also used on the client side, as it is a hydrating
framework. Astro, on the other hand, has a router that is strictly on
the server or at build time: if you build an SPA with Astro you're
basically on your own with the routing.

So in this case SvelteKit is more suitable.

With Kemdict, Astro is more suitable as I don't want hydration there.
The server already finished the routing, and it's not worth it for the
client to replicate that. But in this project we will not have a server.